LayoutTests/fast/images/page-wide-animation-toggle-expected.txt

This test ensures that animation play / pause state is updated when the page-wide 'animations disabled' flag changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Pausing page-wide animation.
PASS internals.isImageAnimating(image) became false
Resuming page-wide image animation.
PASS internals.isImageAnimating(image) became true
Pausing GIF individually.
PASS internals.isImageAnimating(image) became false
Resuming page-wide image animation (this should cause the GIF to begin animating again, despite the fact it was previously paused individually).
PASS internals.isImageAnimating(image) became true
PASS successfullyParsed is true

TEST COMPLETE
